+++ packages/tinacms/src/components/form/FormActions.tsx
@@ -95,7 +95,7 @@ const ActionsOverlay = styled.div<{ open: boolean }>`
   box-shadow: ${shadow('big')};
   background-color: white;
   overflow: hidden;
-  z-index: 100;
+  z-index: 950;
   ${props =>
     props.open &&
     css`