+++ packages/@tinacms/core/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.7.3](https://github.com/tinacms/tinacms/compare/@tinacms/core@0.7.2...@tinacms/core@0.7.3) (2020-03-16)
+
+**Note:** Version bump only for package @tinacms/core
+
+
+
+
+
+## [0.7.2](https://github.com/tinacms/tinacms/compare/@tinacms/core@0.7.2-alpha.0...@tinacms/core@0.7.2) (2020-03-09)
+
+**Note:** Version bump only for package @tinacms/core
+
 ## [0.7.2-alpha.0](https://github.com/tinacms/tinacms/compare/@tinacms/core@0.7.1...@tinacms/core@0.7.2-alpha.0) (2020-03-05)
 
 **Note:** Version bump only for package @tinacms/core