+++ packages/@tinacms/react-core/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.9](https://github.com/tinacms/tinacms/compare/@tinacms/react-core@0.2.8...@tinacms/react-core@0.2.9) (2020-03-16)
+
+**Note:** Version bump only for package @tinacms/react-core
+
+
+
+
+
+## [0.2.8](https://github.com/tinacms/tinacms/compare/@tinacms/react-core@0.2.8-alpha.0...@tinacms/react-core@0.2.8) (2020-03-09)
+
+**Note:** Version bump only for package @tinacms/react-core
+
 ## [0.2.8-alpha.0](https://github.com/tinacms/tinacms/compare/@tinacms/react-core@0.2.7...@tinacms/react-core@0.2.8-alpha.0) (2020-03-05)
 
 **Note:** Version bump only for package @tinacms/react-core