+++ packages/next-tinacms-markdown/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.6](https://github.com/tinacms/tinacms/compare/next-tinacms-markdown@0.2.5...next-tinacms-markdown@0.2.6) (2020-03-16)
+
+**Note:** Version bump only for package next-tinacms-markdown
+
+
+
+
+
+## [0.2.5](https://github.com/tinacms/tinacms/compare/next-tinacms-markdown@0.2.5-alpha.1...next-tinacms-markdown@0.2.5) (2020-03-09)
+
+**Note:** Version bump only for package next-tinacms-markdown
+
 ## [0.2.5-alpha.1](https://github.com/tinacms/tinacms/compare/next-tinacms-markdown@0.2.5-alpha.0...next-tinacms-markdown@0.2.5-alpha.1) (2020-03-06)
 
 **Note:** Version bump only for package next-tinacms-markdown